+++ R/files2SpectraObject.R
@@ -82,8 +82,8 @@
 #' 
 #' If \code{fileExt} contains any of \code{"dx"}, \code{"DX"}, \code{"jdx"} or
 #' \code{"JDX"}, then the files will be processed by \code{\link[readJDX]{readJDX}}.
-#' Consider setting \code{debug = TRUE} for this format, as there are many
-#' options for JCAMP, and many are untested. See \code{\link[readJDX]{readJDX}} for
+#' Consider setting \code{debug = TRUE}, or \code{debug = 1} etc for this format, as there are many
+#' options for JCAMP, and many are untested. See \code{\link[readJDX]{readJDX}} for options and
 #' known limitations.
 #' 
 #' @section matrix2SpectraObject:
@@ -145,8 +145,9 @@
 #' a directory tree containing the data, and/or to specify a starting point other than the current
 #' working directory.  So for instance if the current working directory is not the directory containing
 #' the data files, you can use \code{path = "my_path"} to point to the desired top-level
-#' directory, and \code{recursive = TRUE} to work your way through a set of subdirectories
-#' Also, while argument \code{fileExt} appears to be a file extension (from its
+#' directory, and \code{recursive = TRUE} to work your way through a set of subdirectories.  In addition,
+#' if you are reading in JCAMP-DX files, you can pass arguments to \code{readJDX} via ..., e.g. \code{SOFC = FALSE}.
+#' Finally, while argument \code{fileExt} appears to be a file extension (from its
 #' name and the description elsewhere), it's actually just a grep pattern that you can apply
 #' to any part of the file name if you know how to contruct the proper pattern.
 #'
@@ -164,6 +165,7 @@
 #' 
 #' @importFrom utils read.table setTxtProgressBar txtProgressBar
 #' @importFrom tools file_path_sans_ext
+#' @importFrom readJDX readJDX
 #'
 #' @examples
 #' \dontrun{
@@ -205,12 +207,6 @@ files2SpectraObject <- function(gr.crit = NULL,
 	
 	if (is.null(gr.crit)) stop("No group criteria provided to encode data")
 	
-	# Clean up args found in ... for further use
-	argsLF <- argsRT <- as.list(match.call())[-1] # TWO copies to be used momentarily
-	argsRT <- .cleanArgs(argsRT, "read.table") # further update below
-	argsLF <- .cleanArgs(argsLF, "list.files")
-	argsLF <- c(argsLF, list(pattern = fileExt, full.names = TRUE))
-	
 	DX <- FALSE
 	if (grepl("(dx|DX|jdx|JDX)", fileExt)) {
 		DX <- TRUE
@@ -218,7 +214,14 @@ files2SpectraObject <- function(gr.crit = NULL,
 			stop("You need to install package readJDX to import JCAMP-DX files")
 			}
 		}
-		
+
+	# Clean up args found in ... for further use
+	argsLF <- argsRT <- argsDX <- as.list(match.call())[-1] # THREE copies to be used momentarily
+	argsRT <- .cleanArgs(argsRT, "read.table") # further update below
+	argsLF <- .cleanArgs(argsLF, "list.files")
+	argsLF <- c(argsLF, list(pattern = fileExt, full.names = TRUE))
+	if (DX) argsDX <- .cleanArgs(argsDX, "readJDX")
+	
 	files <- do.call(list.files, argsLF)
 	files.noext <- tools::file_path_sans_ext(basename(files))
 
@@ -231,7 +234,7 @@ files2SpectraObject <- function(gr.crit = NULL,
 		spectra$freq <- temp[,1]
 		}
 	if (DX) {
-		temp <- readJDX::readJDX(file = files[1], debug = debug)
+		temp <- do.call(readJDX::readJDX, args = c(argsDX, list(file = files[1], debug = debug)))
 		spectra$freq <- temp[[4]]$x
 		}
 		
@@ -263,7 +266,7 @@ files2SpectraObject <- function(gr.crit = NULL,
 			spectra$data[i,] <- temp[,2]
 			}
 		if (DX) {
-			temp <- readJDX::readJDX(files[i], debug = debug)
+		    temp <- do.call(readJDX::readJDX, args = c(argsDX, list(file = files[i], debug = debug)))
 			spectra$data[i,] <- temp[[4]]$y
 			}